Use `enum_dispatch` crate to simplify `db::Database`
