+++ Cargo.toml
@@ -15,3 +15,4 @@ rustls = "0.20"
 webpki-roots = "0.22.3"
 rustls-native-certs = "0.6"
 serde = { version = "1", features = ["derive"] }
+enum_dispatch = "0.3.8"
